+++ lib/util/validationResponse.ts
@@ -4,6 +4,8 @@
 import * as pointer from "json-pointer"
 import { CommonError } from "./commonError"
 import { StringMap } from "@ts-common/string-map"
+import { FilePosition } from "@ts-common/source-map"
+import { ValidationEntry } from "yasway"
 
 interface ValidationError {
   validationCategory: string
@@ -15,10 +17,8 @@ interface ValidationError {
   message?: string
   jsonref?: string
   "json-path"?: string
-}
-
-interface Warning {
-  readonly code: unknown
+  jsonUrl?: string
+  jsonPosition?: FilePosition
 }
 
 export class ValidateResponse {
@@ -71,11 +71,17 @@ export class ValidateResponse {
         e.jsonref = jsonpath
         e["json-path"] = pointer.unescape(jsonpath)
       }
+      if (error.jsonUrl && error.jsonUrl.length) {
+        e.jsonUrl = error.jsonUrl
+      }
+      if (error.jsonPosition) {
+        e.jsonPosition = error.jsonPosition
+      }
       return e
     })
   }
 
-  public sanitizeWarnings(warnings: Warning[]): Warning[] {
+  public sanitizeWarnings(warnings: ValidationEntry[]): ValidationEntry[] {
     if (!warnings) {
       throw new Error("validationError cannot be null or undefined.")
     }