+++ lib/validators/semanticValidator.ts
@@ -10,6 +10,7 @@ import * as C from "../util/constants"
 import * as util from "util"
 import { CommonError } from "../util/commonError"
 import { keys } from "@ts-common/string-map"
+import { processErrors } from "../util/processErrors"
 
 export interface Result {
   isValid?: unknown
@@ -48,6 +49,7 @@ export class SemanticValidator extends SpecValidator<SemanticValidationResult> {
       if (validationResult) {
         if (validationResult.errors && validationResult.errors.length) {
           this.specValidationResult.validateSpec.isValid = false
+          processErrors(validationResult.errors)
           const e = this.constructErrorObject(
             ErrorCodes.SemanticValidationError,
             `The spec ${this.specPath} has semantic validation errors.`,
@@ -63,12 +65,13 @@ export class SemanticValidator extends SpecValidator<SemanticValidationResult> {
             `The spec ${this.specPath} is semantically valid.`
         }
         if (validationResult.warnings && validationResult.warnings.length > 0) {
-          const warnings = validateResponse.sanitizeWarnings(validationResult.warnings)
-          if (warnings && warnings.length) {
-            this.specValidationResult.validateSpec.warnings = warnings
+          processErrors(validationResult.warnings)
+          validationResult.warnings = validateResponse.sanitizeWarnings(validationResult.warnings)
+          if (validationResult.warnings && validationResult.warnings.length > 0) {
+            this.specValidationResult.validateSpec.warnings = validationResult.warnings
             log.debug(C.Warnings)
             log.debug("--------")
-            log.debug(util.inspect(warnings))
+            log.debug(util.inspect(validationResult.warnings))
           }
         }
       }